apps/worker/pyproject.toml
[project]
name = "meet-worker"
version = "0.0.0"
description = "MojoMeet LiveKit agent worker"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "httpx>=0.28.1",
    "langchain",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=1.2.28",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-mysql>=3.0.0",
    "livekit-agents[openai,silero]>=1.5.2",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
